requirements.txt
streamlit
rembg
google-generativeai
pillow
onnxruntime